Django/requirements.txt
# Django Framework
Django>=4.2,<5.0
django-cors-headers>=4.3.0

# Django REST Framework
djangorestframework>=3.14.0

# ML Dependencies (shared with ML/ folder)
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL in production

# Security & Authentication
python-decouple>=3.8  # For environment variables

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0

# Code Quality
black>=23.0.0
flake8>=6.1.0

# Optional: Excel export functionality
openpyxl>=3.1.0

# Rasa Integration
requests>=2.31.0

# LLM APIs
google-genai>=0.2.0  # New Gemini API (replaces google-generativeai)
openai>=1.12.0  # For Groq API (OpenAI-compatible)
cohere>=4.47
python-dotenv>=1.0.0
